cleanup parquet code
