+++ polars/polars-arrow/Cargo.toml
@@ -19,4 +19,3 @@ thiserror = "^1.0"
 [features]
 strings = []
 compute = ["arrow/compute_cast", "arrow/compute_concatenate"]
-parquet = ["arrow/io_parquet", "arrow/io_parquet_compression"]
