--- polars/polars-arrow/src/io/mod.rs
@@ -1,5 +1,0 @@
-#[cfg(feature = "parquet")]
-mod parquet;
-
-#[cfg(feature = "parquet")]
-pub use parquet::read_parquet;
