+++ polars/polars-io/src/csv.rs
@@ -41,12 +41,13 @@
 //! }
 //! ```
 //!
+use crate::aggregations::ScanAggregation;
 use crate::csv_core::csv::{cast_columns, CoreReader};
 use crate::csv_core::utils::get_reader_bytes;
 use crate::mmap::MmapBytesReader;
 use crate::predicates::PhysicalIoExpr;
 use crate::utils::resolve_homedir;
-use crate::{ScanAggregation, SerReader, SerWriter};
+use crate::{SerReader, SerWriter};
 pub use arrow::io::csv::write;
 use polars_core::prelude::*;
 #[cfg(feature = "temporal")]
