+++ polars/polars-io/src/parquet/mod.rs
@@ -16,7 +16,7 @@
 //!
 pub mod predicates;
 mod read;
-mod read_par;
+mod read_impl;
 mod write;
 
 use super::*;
