+++ polars/polars-io/src/parquet/read_impl.rs
@@ -1,6 +1,7 @@
+use crate::aggregations::{apply_aggregations, ScanAggregation};
 use crate::mmap::{MmapBytesReader, ReaderBytes};
 use crate::parquet::predicates::collect_statistics;
-use crate::predicates::{arrow_schema_to_empty_df, PhysicalIoExpr};
+use crate::predicates::{apply_predicate, arrow_schema_to_empty_df, PhysicalIoExpr};
 use crate::utils::apply_projection;
 use arrow::array::ArrayRef;
 use arrow::io::parquet::read;
@@ -16,6 +17,103 @@ use std::io::Cursor;
 use std::ops::Deref;
 use std::sync::Arc;
 
+pub fn read_parquet<R: MmapBytesReader>(
+    reader: R,
+    limit: usize,
+    projection: Option<&[usize]>,
+    schema: &ArrowSchema,
+    metadata: Option<FileMetaData>,
+    predicate: Option<Arc<dyn PhysicalIoExpr>>,
+    aggregate: Option<&[ScanAggregation]>,
+) -> Result<DataFrame> {
+    let reader = ReaderBytes::from(&reader);
+    let bytes = reader.deref();
+    let mut reader = Cursor::new(bytes);
+
+    let file_metadata = metadata
+        .map(Ok)
+        .unwrap_or_else(|| read::read_metadata(&mut reader))?;
+    let row_group_len = file_metadata.row_groups.len();
+
+    let projection = projection
+        .map(Cow::Borrowed)
+        .unwrap_or_else(|| Cow::Owned((0usize..schema.fields.len()).collect::<Vec<_>>()));
+
+    let mut dfs = Vec::with_capacity(row_group_len);
+
+    let mut buf_1 = Vec::with_capacity(1024);
+    let mut buf_2 = Vec::with_capacity(1024);
+
+    let mut remaining_rows = limit;
+
+    for rg in 0..row_group_len {
+        let md = &file_metadata.row_groups[rg];
+        if let Some(pred) = &predicate {
+            if let Some(pred) = pred.as_stats_evaluator() {
+                if let Some(stats) = collect_statistics(md.columns(), schema)? {
+                    if !pred.should_read(&stats)? {
+                        continue;
+                    }
+                }
+            }
+        }
+        // test we don't read the parquet file if this env var is set
+        #[cfg(debug_assertions)]
+        {
+            assert!(std::env::var("POLARS_PANIC_IF_PARQUET_PARSED").is_err())
+        }
+
+        let columns = projection
+            .clone()
+            .iter()
+            .map(|column_i| {
+                let b1 = std::mem::take(&mut buf_1);
+                let b2 = std::mem::take(&mut buf_2);
+
+                // the get_column_iterator is an iterator of columns, each column contains compressed pages.
+                // get_column_iterator yields `Vec<Vec<CompressedPage>>`:
+                // outer `Vec` is len 1 for primitive types,
+                // inner `Vec` is whatever number of pages the chunk contains.
+                let column_iter =
+                    read::get_column_iterator(&mut reader, &file_metadata, rg, *column_i, None, b1);
+                let fld = &schema.fields[*column_i];
+                let (mut array, b1, b2) = read::column_iter_to_array(column_iter, fld, b2)?;
+
+                if array.len() > remaining_rows {
+                    array = array.slice(0, remaining_rows);
+                }
+
+                buf_1 = b1;
+                buf_2 = b2;
+
+                Series::try_from((fld.name.as_str(), Arc::from(array)))
+            })
+            .collect::<Result<Vec<_>>>()?;
+
+        remaining_rows = file_metadata.row_groups[rg].num_rows() as usize;
+
+        let mut df = DataFrame::new_no_checks(columns);
+
+        apply_predicate(&mut df, predicate.as_deref())?;
+        apply_aggregations(&mut df, aggregate)?;
+
+        dfs.push(df)
+    }
+
+    if dfs.is_empty() {
+        let schema = if let Cow::Borrowed(_) = projection {
+            Cow::Owned(apply_projection(schema, &projection))
+        } else {
+            Cow::Borrowed(schema)
+        };
+        Ok(arrow_schema_to_empty_df(&schema))
+    } else {
+        let mut df = accumulate_dataframes_vertical(dfs.into_iter())?;
+        apply_aggregations(&mut df, aggregate)?;
+        Ok(df.slice(0, limit))
+    }
+}
+
 pub(crate) fn parallel_read<R: MmapBytesReader>(
     reader: R,
     limit: usize,
@@ -23,6 +121,7 @@ pub(crate) fn parallel_read<R: MmapBytesReader>(
     arrow_schema: &ArrowSchema,
     metadata: Option<FileMetaData>,
     predicate: Option<Arc<dyn PhysicalIoExpr>>,
+    aggregate: Option<&[ScanAggregation]>,
 ) -> Result<DataFrame> {
     let reader = ReaderBytes::from(&reader);
     let bytes = reader.deref();
@@ -54,7 +153,7 @@ pub(crate) fn parallel_read<R: MmapBytesReader>(
     let cont_pool = LowContentionPool::<Vec<u8>>::new(pool_size);
 
     for row_group in 0..n_groups {
-        let md = &file_metadata.row_groups[0];
+        let md = &file_metadata.row_groups[row_group];
         if let Some(pred) = &predicate {
             if let Some(pred) = pred.as_stats_evaluator() {
                 if let Some(stats) = collect_statistics(md.columns(), arrow_schema)? {
@@ -118,11 +217,9 @@ pub(crate) fn parallel_read<R: MmapBytesReader>(
                 .collect::<Result<Vec<_>>>()
         })?;
         let mut df = DataFrame::new_no_checks(columns);
-        if let (Some(predicate), false) = (&predicate, df.is_empty()) {
-            let s = predicate.evaluate(&df)?;
-            let mask = s.bool().expect("filter predicates was not of type boolean");
-            df = df.filter(mask)?;
-        }
+        apply_predicate(&mut df, predicate.as_deref())?;
+        apply_aggregations(&mut df, aggregate)?;
+
         dfs.push(df)
     }
 
@@ -134,6 +231,8 @@ pub(crate) fn parallel_read<R: MmapBytesReader>(
         };
         Ok(arrow_schema_to_empty_df(&schema))
     } else {
-        accumulate_dataframes_vertical(dfs.into_iter()).map(|df| df.slice(0, limit))
+        let mut df = accumulate_dataframes_vertical(dfs.into_iter())?;
+        apply_aggregations(&mut df, aggregate)?;
+        Ok(df.slice(0, limit))
     }
 }
