+++ polars/polars-io/src/predicates.rs
@@ -28,3 +28,16 @@ pub(crate) fn arrow_schema_to_empty_df(schema: &ArrowSchema) -> DataFrame {
         .collect();
     DataFrame::new_no_checks(columns)
 }
+
+#[cfg(any(feature = "ipc", feature = "parquet", feature = "json"))]
+pub(crate) fn apply_predicate(
+    df: &mut DataFrame,
+    predicate: Option<&dyn PhysicalIoExpr>,
+) -> Result<()> {
+    if let (Some(predicate), false) = (&predicate, df.is_empty()) {
+        let s = predicate.evaluate(df)?;
+        let mask = s.bool().expect("filter predicates was not of type boolean");
+        *df = df.filter(mask)?;
+    }
+    Ok(())
+}
