+++ polars/polars-lazy/src/physical_plan/executors/scan.rs
@@ -5,8 +5,9 @@ use crate::logical_plan::LpScanOptions;
 #[cfg(feature = "parquet")]
 use crate::logical_plan::ParquetOptions;
 use crate::utils::try_path_to_str;
+use polars_io::aggregations::ScanAggregation;
+use polars_io::csv::CsvEncoding;
 use polars_io::prelude::*;
-use polars_io::{csv::CsvEncoding, ScanAggregation};
 #[cfg(any(feature = "ipc", feature = "parquet"))]
 use std::fs::File;
 use std::mem;
