+++ polars/polars-lazy/src/physical_plan/expressions/mod.rs
@@ -383,6 +383,7 @@ pub trait PhysicalExpr: Send + Sync {
     /// Can take &dyn Statistics and determine of a file should be
     /// read -> `true`
     /// or not -> `false`
+    #[cfg(feature = "parquet")]
     fn as_stats_evaluator(&self) -> Option<&dyn polars_io::predicates::StatsEvaluator> {
         None
     }
@@ -400,6 +401,7 @@ impl PhysicalIoExpr for PhysicalIoHelper {
         self.expr.evaluate(df, &Default::default())
     }
 
+    #[cfg(feature = "parquet")]
     fn as_stats_evaluator(&self) -> Option<&dyn polars_io::predicates::StatsEvaluator> {
         self.expr.as_stats_evaluator()
     }
